requirements.txt
sqlalchemy
camelot-py[base]
opencv-python
PyPDF2<3.0
great_expectations